Домашняя работа на проверку. Выполнены задания 1 и 2.
